connector/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "connector"
version = "0.1.0"
description = "Conversation-to-CRM connector: WeChat/Feishu/Slack -> local SQLite staging for downstream dashboards."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
]

[project.optional-dependencies]
wechat = ["wcfauto"]
dev = ["pytest>=8", "pytest-asyncio>=0.23"]

[project.scripts]
connector = "connector.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["connector*"]

[tool.setuptools.package-data]
"connector.storage" = ["schema.sql"]
